cleaned up structure for rewrite
