+++ README.md
@@ -1,24 +1,5 @@
-# The PersonalGPT Project
+# The PersonalGPT Rewrite
 <img src="https://github.com/praetor29/personalgpt/blob/main/res/logo.png" width=25% height=25%>
-
-> Logo design by goldiereal
-
-The project aims to create a Discord bot that utilizes the OpenAI API to sustain intelligent, high caliber conversation using the [discord.py](https://github.com/Rapptz/discord.py) python library.
-
-## Features
-- Short Term Memory
-- 100% asynchronous
-- Customizable parameters
-- Runs on OpenAI's `gpt-4-turbo` model
-
-### Please Be Aware
-This project is **self-hosted.** It requires an active OpenAI billing account and Discord API keys.
-
-> [!WARNING]
-> `cognition.py` is severely out of date with the latest OpenAI python implementation. Will be looking to rewrite large chunks of it from the ground up. Lots coming up in the pipeline.
-
-> [!NOTE]
-> While the code is available on the repository, a few files are missing (data folder and prompts) for privacy reasons during testing. The code governing this needs to be reworked to facilitate easier use. This `README` also lacks an installation guide for anyone wishing to replicate.
 
 ## Contributors
 ### Developer
@@ -27,7 +8,7 @@ praetor29
 ```
 ### Beta Testers
 ```plaintext
-yumeshu
-goldiereal
 j1nzou
+goldiereal
+yumeshu
 ```
