--- requirements.txt
@@ -1,7 +1,0 @@
-connectors==4.1
-discord.py==2.3.1
-numpy==1.24.4
-openai==0.27.8
-python-dotenv==1.0.0
-pytz==2023.3
-tiktoken==0.4.0
