--- res/ai_art/dalle3_1.txt
@@ -1,1 +1,0 @@
-Animation Style: A young woman in her late teens with a fair complexion and intensely expressive, bright blue eyes. She sports short blonde hair that reaches below her shoulders, wavy at the ends. She's adorned in a form-fitting, full-sleeved top and pants, contrasting in black and white. Blue and pink accents highlight the cuffs, collar, and seams of her attire, adding a vibrant flair. Her ensemble includes a hood, white on the outside, that she is wearing. She is sitting in a warm, cozy, empty coffee shop with a holographic screen in front of her.
