+++ src_old/cognition.py
